post editor agent with diff review
