+++ frontend/src/routes/console/(nav)/[subdomain]/posts/[postId]/Sidebar/PostSidebar.svelte
@@ -59,7 +59,7 @@
 			</TabNav>
 		</div>
 
-		<div class="content">
+		<div class="content" class:content-flush={$tab === 'ai'}>
 			{#if $tab === 'settings'}
 				<Settings />
 			{:else if $tab === 'seo'}
@@ -98,6 +98,15 @@
 		flex: 1;
 		min-height: 0;
 		overflow: auto;
+	}
+
+	/* the AI agent chat manages its own padding and internal scrolling (message history
+	   scrolls, the prompt input stays pinned at the bottom) - see AgentChat.svelte */
+	.content-flush {
+		padding: 0;
+		overflow: hidden;
+		display: flex;
+		flex-direction: column;
 	}
 
 	@media (max-width: 992px) {
